Uploaded files from local
Contains the two REST API (in python) for extracting insights from taxi trips dataset - 
1. Big Query Client API
2. HTTP GET Request API

Architecture Of Solution:
1. Google Cloud Platform (GCP) - Big Query
2. Python 

Implementation:
Big Query Client API is used to create and integrate Rest API in Python. I implemented two solutions with this approach - 
1) one works directly with the client query (BigQueryClientAPI.py) and
2) the other works with the HTTP Get Request (HttpGetRequest.py). 
Due to the limitations on the usage of GCP services - requires billable service accounts to integrate cloud services, I'm unable to show you the exact response on the fly. The solutions work fine if API is enabled in GCP and a billable account is linked to this project.

Results:
The above approach uses google cloud's public dataset, analyses the data and shares it's response in ~16 seconds.